+++ core/pm2.dev.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  apps: [
+    {
+      name: 'core',
+      script: './src/index.mjs',
+      cwd: '/morio/core',
+      max_memory_restart: '250M',
+      watch: ['./src'],
+    },
+  ],
+}