+++ core/src/lib/cluster.mjs
@@ -2,7 +2,7 @@
 import { testUrl } from '#shared/network'
 import { attempt } from '#shared/utils'
 import { serviceCodes } from '#shared/errors'
-import { serviceOrder, ephemeralServiceOrder, optionalServices } from '#config'
+import { serviceOrder, ephemeralServiceOrder } from '#config'
 // Core imports
 import { ensureMorioNetwork, runHook } from './services/index.mjs'
 import { isBrokerLeading } from './services/broker.mjs'
@@ -55,8 +55,7 @@ async function updateNodeState() {
    */
   const promises = []
   for (const service of utils.isEphemeral() ? ephemeralServiceOrder : serviceOrder) {
-    if (optionalServices.includes(service) || (await runHook('wanted', service)))
-      promises.push(runHook('heartbeat', service))
+    if (await runHook('wanted', service)) promises.push(runHook('heartbeat', service))
   }
   /*
    * Do the same for core as the final service
@@ -125,12 +124,19 @@ export async function ensureMorioClusterConsensus() {
  */
 async function ensureClusterHeartbeat() {
   /*
-   * If we are leading the cluster, don't bother
+   * If we are leading the cluster, run a single heartbeat broadcast.
+   * As leader, we do not have to send heartbeats, but it is possible
+   * that the config was updated on this node, so we need to broadcast
+   * the reload event.
    */
-  if (utils.isLeading()) return false
+  if (utils.isLeading()) {
+    log.debug(`Sending broadcast cluster heartbeat`)
+    runHeartbeat(true, true)
+    return false
+  }
 
   /*
-   * Let people know w're staring the heartbeat
+   * Let people know we're starting the heartbeat
    */
   log.debug(`Starting cluster heartbeat`)
   runHeartbeat(true, false)
@@ -149,7 +155,7 @@ export async function runHeartbeat(broadcast = false, justOnce = false) {
    *
    * Ensure we are comparing to up to date cluster state
    * Unless this is the initial setup in which case we just updated the state
-   * and should perhaps let the world knoww we just work up
+   * and should perhaps let the world know we just work up
    */
   if (!broadcast) await updateClusterState()
 
@@ -204,14 +210,14 @@ export async function runHeartbeat(broadcast = false, justOnce = false) {
 /**
  * Start a local heartbeat
  *
- * When Morio has only 1 node. Or when Morio has only 1 broker node,
+ * When Morio has only 1 node, or when Morio has only 1 broker node,
  * we will run a local heartbeat. This will not reach out over the network
  * but merely trigger the heartbeat lifecycle event locally, as that is what
  * used to keep things up to date.
  *
  * The reason we also run it when there is only 1 broker node is that
  * in a scenario where we have 1 broker node + 1 flanking node, the flanking
- * node going down would mean there is no long a heartbeat, and thus the cluster
+ * node going down would mean there is no longer a heartbeat, and thus the cluster
  * will start to decay.
  *
  */
@@ -263,10 +269,13 @@ async function sendHeartbeat(fqdn, broadcast = false, justOnce = false) {
         },
         to: fqdn,
         cluster: utils.getClusterUuid(),
-        cluster_leader: {
-          serial: utils.getLeaderSerial() || undefined,
-          uuid: utils.getLeaderUuid() || undefined,
-        },
+        // Flanking nodes do not send/know the cluster leader
+        cluster_leader: utils.isFlankingNode()
+          ? undefined
+          : {
+              serial: utils.getLeaderSerial() || undefined,
+              uuid: utils.getLeaderUuid() || undefined,
+            },
         version: utils.getVersion(),
         settings_serial: Number(utils.getSettingsSerial()),
         keys_serial: Number(utils.getKeysSerial()),
@@ -304,7 +313,7 @@ async function sendHeartbeat(fqdn, broadcast = false, justOnce = false) {
   verifyHeartbeatResponse({ fqdn, data, rtt })
 
   /*
-   * Trigger a new heatbeat
+   * Trigger a new heartbeat
    */
   if (!justOnce) runHeartbeat(false, false)
 }
@@ -340,7 +349,7 @@ function verifyHeartbeatResponse({ fqdn, data, rtt = 0, error = false }) {
    */
   if (error || data?.code) {
     /*
-     * Storing the result of a failed hearbteat will influence the cluster state
+     * Storing the result of a failed heartbeat will influence the cluster state
      */
     utils.setHeartbeatIn(fqdn, { up: false, ok: false, error: error.code })
     /*
@@ -405,7 +414,7 @@ function verifyHeartbeatResponse({ fqdn, data, rtt = 0, error = false }) {
   } else if (Array.isArray(data?.nodes)) {
     for (const uuid in data.nodes) {
       /*
-       * It it's a valid hearbeat, add the node info to the state
+       * If it's a valid hearbeat, add the node info to the state
        */
       if (uuid !== utils.getNodeUuid()) utils.setClusterNode(uuid, data.nodes[uuid])
     }
@@ -460,7 +469,7 @@ export async function verifyHeartbeatRequest(data, type = 'heartbeat') {
   }
 
   /*
-   * Verify the 'to' is really us as a mismatch here can indicate fault DNS configuration
+   * Verify the 'to' is really us as a mismatch here can indicate faulty DNS configuration
    */
   if (utils.getNodeFqdn() !== data.to) {
     const err = 'HEARTBEAT_TARGET_FQDN_MISMATCH'
@@ -502,8 +511,11 @@ export async function verifyHeartbeatRequest(data, type = 'heartbeat') {
   /*
    * Verify leader (only for heatbeats)
    * If there's a mismatch, ask to re-elect the cluster leader.
+   * However, flanking nodes do not know/care who the cluster leader is.
+   * So this only matters for broker nodes.
+   * Note that a healthcheck from a flanking node will not include data.cluster_leader.
    */
-  if (!data.cluster_leader?.serial || data.cluster_leader.serial !== utils.getLeaderSerial()) {
+  if (data.cluster_leader && data.cluster_leader.serial !== utils.getLeaderSerial()) {
     /*
      * Do they look to us as their leader?
      */
@@ -573,7 +585,7 @@ export async function verifyHeartbeatRequest(data, type = 'heartbeat') {
  *
  * This is called from the beforeall lifecycle hook
  * Note that Morio always runs in cluster mode
- * to ensure we can reach flanking nodes whne they are added.
+ * to ensure we can reach flanking nodes when they are added.
  */
 export async function ensureMorioCluster() {
   utils.setCoreReady(false)
@@ -615,7 +627,7 @@ export async function ensureMorioCluster() {
 }
 
 /*
- * Helpoer method to invite a single node to join the cluster
+ * Helper method to invite a single node to join the cluster
  *
  * @param {string} fqdn - The fqdn of the remote node
  */
@@ -655,7 +667,7 @@ async function inviteClusterNodeAttempt(remote) {
   const flanking = utils.isThisAFlankingNode({ fqdn: remote })
 
   /*
-   * Load data from disk becauise what we sync between cluster nodes
+   * Load data from disk because what we sync between cluster nodes
    * is what is written to disk.
    */
   const timestamp = utils.getSettingsSerial()
@@ -690,6 +702,7 @@ async function inviteClusterNodeAttempt(remote) {
   /*
    * Validate response
    */
+  if (result.status && result.status !== 200) return false
   const [valid, err] = await validate(`res.cluster.join`, result)
   if (valid) log.info(`Node ${valid.node} will join the cluster`)
   else log.todo(err, `Handle cluster join failure.`)