+++ ui/lib/utils.mjs
@@ -7,6 +7,14 @@ import _slugify from 'slugify'
 import { jwtDecode } from 'jwt-decode'
 import { roles } from 'config/roles.mjs'
 
+export const arrayMembers = (set, op='add', data) => {
+  const s = new Set(set)
+  if (op === 'add') s.add(data)
+  else s.delete(data)
+
+  return [...s]
+}
+
 export const asJson = (data, pretty = true) => {
   const json = {}
   for (const [key, val] of Object.entries(data)) {